# [opt](topn)Optimize the time for topn to lazy materialized reading of external tables