--- a/be/src/pipeline/dependency.cpp
+++ b/be/src/pipeline/dependency.cpp
@@ -21,6 +21,7 @@
 #include <mutex>
 
 #include "common/logging.h"
+#include "exec/rowid_fetcher.h"
 #include "pipeline/exec/multi_cast_data_streamer.h"
 #include "pipeline/pipeline_fragment_context.h"
 #include "pipeline/pipeline_task.h"
@@ -28,6 +29,7 @@
 #include "runtime/memory/mem_tracker.h"
 #include "runtime_filter/runtime_filter_consumer.h"
 #include "util/brpc_client_cache.h"
+#include "vec/exec/scan/file_scanner.h"
 #include "vec/exprs/vectorized_agg_fn.h"
 #include "vec/exprs/vslot_ref.h"
 #include "vec/spill/spill_stream_manager.h"
@@ -481,6 +483,11 @@ Status MaterializationSharedState::merge_multi_response(vectorized::Block* block
             DCHECK(rpc_struct.callback->response_->blocks_size() > i);
             RETURN_IF_ERROR(
                     partial_block.deserialize(rpc_struct.callback->response_->blocks(i).block()));
+            if (rpc_struct.callback->response_->blocks(i).has_profile()) {
+                auto response_profile = RuntimeProfile::from_proto(
+                        rpc_struct.callback->response_->blocks(i).profile());
+                RETURN_IF_ERROR(_update_profile_info(backend_id, response_profile.get()));
+            }
 
             if (!partial_block.is_empty_column()) {
                 _block_maps[backend_id] = std::make_pair(std::move(partial_block), 0);
@@ -531,6 +538,36 @@ Status MaterializationSharedState::merge_multi_response(vectorized::Block* block
     }
     origin_block.clear();
 
+    return Status::OK();
+}
+
+Status MaterializationSharedState::_update_profile_info(int64_t backend_id,
+                                                        RuntimeProfile* response_profile) {
+    if (!backend_profile_info_string.contains(backend_id)) {
+        backend_profile_info_string.emplace(backend_id,
+                                            std::map<std::string, std::stringstream> {});
+    }
+    auto& info_map = backend_profile_info_string[backend_id];
+
+    auto func = [&](const std::string& info_key) -> Status {
+        const auto* info_value = response_profile->get_info_string(info_key);
+        if (info_value == nullptr) [[unlikely]] {
+            LOG(WARNING) << "Get row id fetch rpc profile success, but no info key :" << info_key;
+            return Status::OK();
+        }
+        if (!info_map.contains(info_key)) {
+            info_map.emplace(info_key, std::stringstream {});
+        }
+        info_map[info_key] << *info_value << ", ";
+        return Status::OK();
+    };
+
+    RETURN_IF_ERROR(func(RowIdStorageReader::ScannersRunningTimeProfile));
+    RETURN_IF_ERROR(func(RowIdStorageReader::InitReaderAvgTimeProfile));
+    RETURN_IF_ERROR(func(RowIdStorageReader::GetBlockAvgTimeProfile));
+    RETURN_IF_ERROR(func(RowIdStorageReader::FileReadLinesProfile));
+    RETURN_IF_ERROR(func(vectorized::FileScanner::FileReadBytesProfile));
+    RETURN_IF_ERROR(func(vectorized::FileScanner::FileReadTimeProfile));
     return Status::OK();
 }
 
@@ -657,6 +694,12 @@ Status MaterializationSharedState::init_multi_requests(
                                                              .request = multi_get_request,
                                                              .callback = nullptr,
                                                              .rpc_timer = MonotonicStopWatch()});
+        //        if (profile != nullptr) {
+        //            profile.add_co
+        //
+        //        }
+        //        profile->add_child()
+        //        node_info.printTo()
     }
     // add be_num ad count finish counter for source dependency
     ((CountedFinishDependency*)source_deps.back().get())->add((int)rpc_struct_map.size());