--- a/be/src/pipeline/dependency.h
+++ b/be/src/pipeline/dependency.h
@@ -828,6 +828,10 @@ public:
 
     void create_counter_dependency(int operator_id, int node_id, const std::string& name);
 
+private:
+    Status _update_profile_info(int64_t backend_id, RuntimeProfile* response_profile);
+
+public:
     bool rpc_struct_inited = false;
     AtomicStatus rpc_status;
 
@@ -842,6 +846,8 @@ public:
     // Register each line in which block to ensure the order of the result.
     // Zero means NULL value.
     std::vector<std::vector<int64_t>> block_order_results;
+    // backend id => <rpc profile info string key, rpc profile info string value>.
+    std::map<int64_t, std::map<std::string, std::stringstream>> backend_profile_info_string;
 };
 #include "common/compile_check_end.h"
 } // namespace doris::pipeline