### What problem does this PR solve?
Related PR: #51329

Problem Summary:
Topn lazy materialize was introduced in pr#51329 , but the implementation had performance issues when reading external tables. This pr is used for optimization.
1. Before this, the materialization phase read one row of data from the file each time. This pr classifies according to scan_range and reads multiple rows of data from the file at one time.
2. Before this, the materialization phase was a single-threaded file reading phase. This pr creates a scan task and submits the task to the workload group to improve the reading speed.
3. Before this, the runtime profile was transmitted through thrift. This pr introduces the implementation of protobuf and adds the profile information of `RowIDFetcher` to `MATERIALIZATION_OPERATOR`. 
The example is as follows:
1FE 2BE
sql :select * from ali_hive.tpch100_orc.lineitem order by l_partkey limit 10;
```
MATERIALIZATION_OPERATOR  (id=3):(ExecTime:  2.645ms)
        -  BlocksProduced:  5
        -  CloseTime:  0ns
        -  ExecTime:  2.645ms
        -  InitTime:  0ns
        -  MemoryUsage:  0.00  
        -  MemoryUsagePeak:  0.00  
        -  OpenTime:  0ns
        -  ProjectionTime:  528.913us
        -  RowsProduced:  10
        -  WaitForDependency[MATERIALIZATION_COUNTER_DEPENDENCY]Time:  12sec874ms
    RowIDFetcher:  BackendId:1750838859134:
            -  FileReadBytes:  {[2.89  MB,  ],  [9.51  MB,  ],  [6.81  MB,  ],  [4.74  MB,  ],  [22.33  MB,  ],  }
            -  FileReadLines:  {[1,  ],  [1,  ],  [1,  ],  [1,  ],  [1,  ],  }
            -  FileReadTime:  {[102.960ms,],  [104.028ms,],  [99.817ms,],  [98.260ms,],  [120.129ms,],  }
            -  GetBlockAvgTime:  {14ms,  2ms,  2ms,  1ms,  3ms,  }
            -  InitReaderAvgTime:  {14ms,  2ms,  2ms,  1ms,  3ms,  }
            -  ScannersRunningTime:  {130ms,  124ms,  116ms,  113ms,  151ms,  }
    RowIDFetcher:  BackendId:1750936290862:
            -  FileReadBytes:  {[13.80  MB,  ],  [21.28  MB,  ],  [8.18  MB,  ],  [16.69  MB,  ],  [19.16  MB,  ],  }
            -  FileReadLines:  {[1,  ],  [1,  ],  [1,  ],  [1,  ],  [1,  ],  }
            -  FileReadTime:  {[113.031ms,],  [132.087ms,],  [105.361ms,],  [117.245ms,],  [125.535ms,],  }
            -  GetBlockAvgTime:  {2ms,  2ms,  2ms,  1ms,  3ms,  }
            -  InitReaderAvgTime:  {2ms,  2ms,  2ms,  1ms,  3ms,  }
            -  ScannersRunningTime:  {144ms,  160ms,  127ms,  142ms,  159ms,  }
```


### Release note

None

### Check List (For Author)

- Test <!-- At least one of them must be included. -->
    - [ ] Regression test
    - [ ] Unit Test
    - [ ] Manual test (add detailed scripts or steps below)
    - [ ] No need to test or manual test. Explain why:
        - [ ] This is a refactor/code format and no logic has been changed.
        - [ ] Previous test can cover this change.
        - [ ] No code files have been changed.
        - [ ] Other reason <!-- Add your reason?  -->

- Behavior changed:
    - [ ] No.
    - [ ] Yes. <!-- Explain the behavior change -->

- Does this need documentation?
    - [ ] No.
    - [ ] Yes. <!-- Add document PR link here. eg: https://github.com/apache/doris-website/pull/1214 -->

### Check List (For Reviewer who merge this PR)

- [ ] Confirm the release note
- [ ] Confirm test cases
- [ ] Confirm document
- [ ] Add branch pick label <!-- Add branch pick label that this PR should merge into -->