@hubgeter 2025-06-22
run buildall