@hubgeter 2025-06-24
run buildall