@hello-stephen 2025-06-25
# BE Regression && UT Coverage Report
Increment line coverage `11.19% (16/143)` :tada:

[Increment coverage report](http://coverage.selectdb-in.cc/coverage/52114_0a38597368c5347fa335644d05cd4c8d8f8c86cd_merge/increment_report/index.html)
[Complete coverage report](http://coverage.selectdb-in.cc/coverage/52114_0a38597368c5347fa335644d05cd4c8d8f8c86cd_merge/report/index.html)
| Category          | Coverage           |
|-------------------|--------------------|
| Function Coverage | 69.53% (18419/26492) |
| Line Coverage     | 61.55% (185937/302107) |
| Region Coverage   | 60.14% (109657/182341) |
| Branch Coverage   | 53.75% (56740/105554) |