@hubgeter 2025-07-01
run buildall