@hubgeter 2025-07-07
run buildall