@hubgeter 2025-07-08
run buildall